backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0

# Additional ML and data processing
matplotlib==3.8.2
seaborn==0.13.0
imbalanced-learn==0.11.0

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
